Done the clustering now returns a City object as Json with a topFivePhotos array.Also done the testing.

+++ src/main/java/com/cityGuideTL/backend/Services/HaversineDistanceService.java
@@ -7,11 +7,10 @@ public class HaversineDistanceService {
 
     public static double distance(Point startPoint, Point endPoint) {
 
-        double startLat = startPoint.getLatitude().doubleValue();
-        double startLong = startPoint.getLongitude().doubleValue();
-        double endLat = endPoint.getLatitude().doubleValue();
-        double endLong = endPoint.getLongitude().doubleValue();
-
+        double startLat = startPoint.getLatitude();
+        double startLong = startPoint.getLongitude();
+        double endLat = endPoint.getLatitude();
+        double endLong = endPoint.getLongitude();
 
         double dLat  = Math.toRadians((endLat - startLat));
         double dLong = Math.toRadians((endLong - startLong));
@@ -27,7 +26,5 @@ public class HaversineDistanceService {
 
     public static double haversine(double val) {
         return Math.pow(Math.sin(val / 2), 2);
-
     }
-
 }
